+++ resources/views/back/_fields/input.blade.php
@@ -1,12 +1,12 @@
 @if($field->prepend)
     <div class="input-addon">
-        <div class="addon inset-s">{!! $field->prepend !!}</div>
+        <div class="addon inset-s">{!! $field->translateValue($field->prepend, $locale ?? null) !!}</div>
         <input type="text" name="{{ $name ?? $key }}" id="{{ $key }}" class="input inset-s" placeholder="{{ $field->placeholder ?? '' }}" value="{{ old($key, $manager->getFieldValue($key)) }}">
     </div>
 @elseif($field->append)
     <div class="input-addon">
         <input type="text" name="{{ $name ?? $key }}" id="{{ $key }}" class="input inset-s" placeholder="{{ $field->placeholder ?? '' }}" value="{{ old($key, $manager->getFieldValue($key)) }}">
-        <div class="addon inset-s">{!! $field->append !!}</div>
+        <div class="addon inset-s">{!! $field->translateValue($field->append, $locale ?? null) !!}</div>
     </div>
 @else
     <input type="text" name="{{ $name ?? $key }}" id="{{ $key }}" class="input inset-s" placeholder="{{ $field->placeholder ?? '' }}" value="{{ old($key, $manager->getFieldValue($key)) }}">