+++ src/Menu/ActsAsMenuItem.php
@@ -4,7 +4,7 @@ namespace Thinktomorrow\Chief\Menu;
 
 interface ActsAsMenuItem
 {
-    public function menuUrl(): string;
+    public function url(): string;
 
     public function menuLabel(): string;
 }