+++ tests/Feature/Menu/CreateMenuItemTest.php
@@ -19,16 +19,17 @@ class CreateMenuItemTest extends TestCase
     /** @test */
     public function admin_can_view_the_create_form()
     {
-        $this->disableExceptionHandling();
         $response = $this->asAdminWithoutRole()->get(route('chief.back.menuitem.create', 'main'));
-        $response->assertStatus(200);
+        $response->assertViewIs('chief::back.menu.create')
+                 ->assertStatus(200);
     }
 
     /** @test */
     public function guests_cannot_view_the_create_form()
     {
         $response = $this->get(route('chief.back.menuitem.create', 'main'));
-        $response->assertStatus(302)->assertRedirect(route('chief.back.login'));
+        $response->assertStatus(302)
+                 ->assertRedirect(route('chief.back.login'));
     }
 
     /** @test */
@@ -140,7 +141,7 @@ class CreateMenuItemTest extends TestCase
                 'trans.nl.url'      => 'thinktomorrow.be',
             ]));
 
-        $this->assertEquals('http://thinktomorrow.be', MenuItem::first()->url);
+        $this->assertEquals('https://thinktomorrow.be', MenuItem::first()->url);
     }
 
     /** @test */
@@ -212,7 +213,7 @@ class CreateMenuItemTest extends TestCase
         $this->assertEquals($overrides['parent_id'] ?? null, $menuItem->parent_id);
 
         $this->assertEquals($overrides['trans.nl.label'] ?? 'nieuw label', $menuItem->{'label:nl'});
-        $this->assertEquals($overrides['trans.nl.url'] ?? 'http://google.com', $menuItem->{'url:nl'});
+        $this->assertEquals($overrides['trans.nl.url'] ?? 'https://google.com', $menuItem->{'url:nl'});
 
         $this->assertEquals($overrides['trans.en.label'] ?? '', $menuItem->{'label:en'});
         $this->assertEquals($overrides['trans.en.url'] ?? '', $menuItem->{'url:en'});