+++ tests/Feature/PageBuilder/PagetitleTest.php
@@ -7,7 +7,6 @@ use Thinktomorrow\Chief\Management\Register;
 use Thinktomorrow\Chief\Modules\Module;
 use Thinktomorrow\Chief\Modules\PagetitleModule;
 use Thinktomorrow\Chief\Modules\TextModule;
-use Thinktomorrow\Chief\Pages\Application\CreatePage;
 use Thinktomorrow\Chief\Pages\PageManager;
 use Thinktomorrow\Chief\Tests\Fakes\ArticlePageFake;
 use Thinktomorrow\Chief\Tests\TestCase;
@@ -23,6 +22,7 @@ class PagetitleTest extends TestCase
         parent::setUp();
 
         $this->setUpDefaultAuthorization();
+        config()->set('app.fallback_locale', 'nl');
 
         app(Register::class)->register('articles', PageManager::class, ArticlePageFake::class);