+++ tests/fakes/ProductPageFake.php
@@ -8,4 +8,5 @@ use Thinktomorrow\Chief\Pages\Page;
 
 class ProductPageFake extends Page
 {
+    protected static $baseUrlSegment = 'products';
 }